refactor: Extract session_opts_for_workflow into AI.SessionConfig
Move session_opts_for_workflow/3 and merge_phase_opts/2 from ClaudeSession
(infrastructure) into a new SessionConfig module, removing upward dependencies
on Destila.Workflows and Destila.AI from the GenServer.
